content/xml/document/test/test_bug445330.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=445330
-->
<head>
  <title>Test for Bug 445330</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=445330">Mozilla Bug 445330</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 445330 **/

var loadCounter = 0;
var dclCounter = 0;

function runTest() {
  var doc = document.implementation.createDocument("", "", null);
  doc.onload = function(evt) {
      ++loadCounter;
      ok(dclCounter == 1, "DOMContentLoaded should have fired.");
      if (loadCounter == 1 && dclCounter == 1) {
        SimpleTest.finish();
      }
    };

  doc.addEventListener("DOMContentLoaded",
    function (evt) {
      ++dclCounter;
      ok(loadCounter == 0, "load event shouldn't have fired.");
      if (loadCounter == 1 && dclCounter == 1) {
        SimpleTest.finish();
      }
    },
    true);

  doc.load("");
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
